Adds HL keybinding for cursor movement
Replaces the boolean setting for JK cursor movement with an enum to support additional keybindings.

This allows users to configure cursor movement when the candidate window is open to use either the JK keys or the HL keys, providing more flexibility. The setting is updated in both ChromeOS and Pime implementations.